Title: 자연수 n개의 합, Time: 115ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 250407/자연수 n개의 합/sum-of-n-natural-numbers.java
@@ -6,7 +6,7 @@ public class Main {
 	public static Scanner sc = new Scanner(System.in);
 
 	public static void input() {
-		s = sc.nextInt();
+		s = sc.nextLong();
 	}
 
 	// 합이 S이하인 경우 중 N의 최대 값
